+++ routes/api/usersApi.js
@@ -2,26 +2,7 @@ const express = require("express");
 const router = express.Router();
 const ctrlUser = require("../../controller/usersController");
 const authorization = require("../../tools/authorization");
-
-const path = require("path");
-const multer = require("multer");
-const uploadDir = path.join(process.cwd(), "tmp");
-
-const storage = multer.diskStorage({
-  destination: (_, __, cb) => {
-    cb(null, uploadDir);
-  },
-  filename: (_, file, cb) => {
-    cb(null, file.originalname);
-  },
-  limits: {
-    fileSize: 1048576,
-  },
-});
-
-const upload = multer({
-  storage: storage,
-});
+const uploadPicture = require("../../tools/uploadPicture");
 
 router.post("/signup", ctrlUser.signUp);
 router.post("/login", ctrlUser.logIn);
@@ -31,7 +12,7 @@ router.patch("/", authorization, ctrlUser.changeSubscription);
 router.patch(
   "/avatars",
   authorization,
-  upload.single("avatar"),
+  uploadPicture("avatar"),
   ctrlUser.avatar
 );
 router.get("/verify/:verificationToken", ctrlUser.emailVerification)
